# api-docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.APIKeysWrapper:
    properties:
      api_keys:
        items:
          $ref: '#/definitions/api.APIKeyListResponse'
        type: array
    type: object
  api.ChangePasswordRequest:
    properties:
      confirmPassword:
        type: string
      currentPassword:
        type: string
      newPassword:
        minLength: 6
        type: string
    required:
    - confirmPassword
    - currentPassword
    - newPassword
    type: object
  api.ChangeUsernameRequest:
    properties:
      newUsername:
        maxLength: 50
        minLength: 3
        type: string
      password:
        type: string
    required:
    - newUsername
    - password
    type: object
...
  api.CreateAPIKeyRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  api.CreateAPIKeyResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
    type: object
...
  api.RefreshTokenResponse:
    properties:
      token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Audio transcription service using WhisperX
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: aScribe API
  version: "1.0"
paths:
...
  /api/v1/api-keys:
    get:
      description: Get all API keys for the current user (without exposing the actual
        keys)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIKeysWrapper'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Create a new API key for external API access
      parameters:
      - description: API key creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - api-keys
  /api/v1/api-keys/{id}:
    delete:
      description: Delete an API key
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - api-keys
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - auth
  /api/v1/auth/change-username:
    post:
      consumes:
      - application/json
      description: Change the current user's username
      parameters:
      - description: Username change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChangeUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change username
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Logout user and invalidate token (client-side action)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      description: Rotate refresh token and return new access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RefreshTokenResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register the initial admin user (only allowed when no users exist)
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register initial admin user
      tags:
      - auth
...
  /api/v1/chat/models:
    get:
      description: Get list of available OpenAI chat models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChatModelsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get available chat models
      tags:
      - chat
  /api/v1/chat/sessions:
    post:
      consumes:
      - application/json
      description: Create a new chat session for a transcription
      parameters:
      - description: Chat session creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChatCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.ChatSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create a new chat session
      tags:
      - chat
  /api/v1/chat/sessions/{session_id}:
    delete:
      description: Delete a chat session and all its messages
      parameters:
      - description: Chat Session ID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete a chat session
      tags:
      - chat
    get:
      description: Get a specific chat session with all its messages
      parameters:
      - description: Chat Session ID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChatSessionWithMessages'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get a chat session with messages
      tags:
      - chat
  /api/v1/chat/sessions/{session_id}/messages:
    post:
      consumes:
      - application/json
      description: Send a message to a chat session and get streaming response
      parameters:
      - description: Chat Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/api.ChatMessageRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Streaming response
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Send a message to a chat session
      tags:
      - chat
  /api/v1/chat/sessions/{session_id}/title:
    put:
      consumes:
      - application/json
      description: Update the title of a chat session
      parameters:
      - description: Chat Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: Title update request
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChatSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update chat session title
      tags:
      - chat
  /api/v1/chat/sessions/{session_id}/title/auto:
    post:
      description: Uses the configured LLM to summarize the first exchange into a
        concise title. Only updates if the current title appears default/user-unset.
      parameters:
      - description: Chat Session ID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChatSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Auto-generate chat session title
      tags:
      - chat
  /api/v1/chat/transcriptions/{transcription_id}/sessions:
    get:
      description: Get all chat sessions for a specific transcription
      parameters:
      - description: Transcription ID
        in: path
        name: transcription_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ChatSessionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get chat sessions for a transcription
      tags:
      - chat
  /api/v1/config/openai/validate:
    post:
      consumes:
      - application/json
      description: Validate the provided OpenAI API key and return available Whisper
        models
      parameters:
      - description: API Key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ValidateOpenAIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Validate OpenAI API Key
      tags:
      - config
...
  /api/v1/llm/config:
    get:
      description: Get the current active LLM configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LLMConfigResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get LLM configuration
      tags:
      - llm
    post:
      consumes:
      - application/json
      description: Create or update LLM configuration settings
      parameters:
      - description: LLM configuration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.LLMConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LLMConfigResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create or update LLM configuration
      tags:
      - llm
  /api/v1/notes/{note_id}:
    delete:
      description: Delete a note by its ID
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a note
      tags:
      - notes
    get:
      description: Get a note by its ID
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get a note
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Update the content of a note
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      - description: Note update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.NoteUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update a note
      tags:
      - notes
  /api/v1/profiles:
    get:
      description: Get list of all transcription profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TranscriptionProfile'
            type: array
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List transcription profiles
      tags:
      - profiles
    post:
      consumes:
      - application/json
      description: Create a new transcription profile
      parameters:
      - description: Profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.TranscriptionProfile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TranscriptionProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create transcription profile
      tags:
      - profiles
  /api/v1/profiles/{id}:
    delete:
      description: Delete a transcription profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete transcription profile
      tags:
      - profiles
    get:
      description: Get a transcription profile by ID
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TranscriptionProfile'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get transcription profile
      tags:
      - profiles
    put:
      consumes:
      - application/json
      description: Update a transcription profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.TranscriptionProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TranscriptionProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update transcription profile
      tags:
      - profiles
  /api/v1/profiles/{id}/set-default:
    post:
      description: Mark the specified profile as the default profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Set default transcription profile
      tags:
      - profiles
...
  /api/v1/transcription/models:
    get:
      description: Get list of supported WhisperX models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get supported models
      tags:
      - transcription
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"